remove old file and assertEquals once again
